Separate mfhd_item from item_vw to fix encoding problem
The table item_vw is encoded in UTF-8 except for the mfhd_item
part that is ISO-8859-1. In config.pl it has been specified that
the item_vw is UTF-8 so now that mfhd_item is in mhfd_item2 instead
of item_vw table we use the correct (default) ISO-8859-1 encoding.